refactoring profile, setting
